2024 day 6
I tried to game the system in my initial implementation of part 1, assuming part 2 would be something like "what if the guard sometimes turns left rather than right".  This final version is a simplified algorithm for part 1, which we call recursively to solve part 2.

Technically my solution allows for the part 1 path to be a loop as well, but the problem is still well defined in that case.