2024 day 11 - day 11 shouldn't be this easy.
I just did the trivial memoized recursive implementation for part 1, just to *count* the number of resulting stones, and it ran in a few ms.  I expected part 2 to make some use of the "order doesn't change" thing, but in the end it was literally a one character change from part 1 and still ran in milliseconds.

Thank you @functools.cache