2024 day 21
I guess I was conditioned by shortest-path-finding from all the previous problems that required it, so my first attempt was a "forwards" approach modelling the whole state space and finding shortest paths from start to end.  This was fine for part 1 but way to slow for part 2, so I tried again with a "backwards" approach, starting from the end goal and working out what steps the previous robot in the chain has to make to achieve that goal.  This was many magnitudes faster.