2025 day 9
Part 1: naive algorithm to just check the area of all possible rectangles and pick the biggest.

Part 2: re-used the logic I originally developed in 2023 day 18, which was another similar construction where you have a list of points that are corners of a biiiiig polygon where all the angles are 90 degrees.  To make that problem tractable I essentially did a kind of run-length encoding compression on the coordinate space, as we only really care about the corners and the vertical and horizontal edges of the polygon, not the huge empty spaces in between.